Refactor the locally embedded document field
It was duplicated
